sys_python/requirements.txt
flask==3.1.3
flask_cors==6.0.2
qqmusic-api-python==0.5.2
httpx==0.28.1
